+++ src/options/functions.ts
@@ -1,0 +1,4 @@
+export function clone(source: HTMLTemplateElement): HTMLElement
+{
+	return (source.content.cloneNode(true) as DocumentFragment).firstElementChild as HTMLElement;
+}